+++ sdk/identity/identity/src/client/identityClient.ts
@@ -90,6 +90,9 @@ export class IdentityClient extends ServiceClient implements INetworkModule {
 
     super({
       requestContentType: "application/json; charset=utf-8",
+      retryOptions: {
+        maxRetries: 3,
+      },
       ...options,
       userAgentOptions: {
         userAgentPrefix,